Write functioning queue.js
Wrote a functioning queue that will delete
the first the oldest queued variable, and
append a value to the end of an storage.

I'd like to refactor granted there are so
many different variables that had to be
declared, and as Kan remarked that this
functional style can be taxing on a memory
bank.