Write shared function instantiation for queue.js
Created constructor function that instantiates an
empty object, places properties on that empty object
and extends the object further with the properties
of a seperate object.

declared properties last and first to help with
pinpointing the first item in the storage and
the last item. I want to think of a way to find
the first and last without having to declare the
properties to free up space in the memory.