Wrote a prototypal instantiation of queue
With the help of Alex Kim, was able create a
queue without the extra properities of 'first'
and 'last' appended as properties of the object.
This saves on memory. This is accomplished by
utilizing the Object.keys() method to determine
size, and with Object.keys searching for the first
item in the array it is referring to.

Another thing to note is that prototypes allow for
the objects of Queue to contain less properties,
and therefore are 'lighter'.  Instead, they have
access to the properties of another object. My mind
is blown. So cool.