Wrote operable graph.js
Discussed with Zach and Hugo a minor bug
that somehow implements prototype methods
before console logging. Will have to look
into the reasons for this behavior.

Also discussed with Hugo and Zach my use
of _.score in the forEachNode Method on
lines 43, and 44.  Discussed the prospect
of changing the tests if the implementation
is has sound logic and reason. Both gave
pros and cons.

Pros: We are not bound to the tests, and
TDD is subject to the dev as a tool to
prevent consequences.  It is a contract
and, on occasion if the contract is unfair,
it is perfectly reasonable to change the
terms of the contract.

However, granted that most software
development is built in teams, the cost
of changing a test when not communicated
well or not fully reasonable can result
in detrimental and costly consequences
down the line for you and other devs.