Provide operable Set.js
By far my favorite data structure. Because it is
easy to make.  However, the drawbacks are of course
there are not connections that can be made for each
node within the set.  The time complexity is 0(1) for
each of the methods it utilizes.  Also to note that
a set can only have unique items within.